+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Version 0.10.0
 --------------
 
-Unreleased
+Released 2023-01-09
 
 - Improve error message when ``FileSystemCache`` methods are called with non-str keys. :pr:`170`
 - Added ``DynamoDb`` as a cache backend :pr:`209`